PoC Work for GraphQL
This is a very lazy git comment, and for that I'm sorry.

This is the start of intregrating GraphQL as an API layer, that wraps the streamlab layer. It's intended to evolve over time, and could then contain data from other sources.

It's using some uncompiled react in the frontend as a PoC.

That's pretty much it. Lots will follow, and lot's will be documented in the future (HAHAHAHAHA).