Split up codebase
This is first step in creating an frontend, and a API. These can be deployed independed of each other.

Start the client folder with create-react-app, this will make us compile the react app correctly, and not in the bowser as before, this makes the client code a bit cleaner, and gives us better options when coding.